+++ src/Commands/InstallCommand.php
@@ -4,7 +4,7 @@ namespace kranack\Lint\Commands;
 
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 use kranack\Lint\Env\Environment;
@@ -18,7 +18,8 @@ class InstallCommand extends Command
     {
 		$this
 			->setDescription('Install config')
-			->setHelp('This command install config for lint');
+			->setHelp('This command install config for lint')
+			->addOption('force', 'f', InputOption::VALUE_NONE, 'Force install');
 	}
 	
 	protected function isInstalled()
@@ -28,12 +29,21 @@ class InstallCommand extends Command
 		}
 	}
 
+	protected function install()
+	{
+		(new Environment())->init();
+	}
+
     protected function execute(InputInterface $input, OutputInterface $output)
     {
+		$force = $input->getOption('force');
+
 		try {
 			$this->isInstalled();
+
+			if ($force) { $this->install(); }
 		} catch (EnvironmentNotConfigured $e) {
-			(new Environment())->init();
+			$this->install();
 		}
     }
 }